requirements.txt
# ReconStruct - 2D Blueprint to 3D Model Conversion
# Image Processing (core; no Python 3.13+ restriction)
opencv-python>=4.8.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
# Optional ML backends (uncomment for TF/PyTorch; TensorFlow needs Python <3.13):
# tensorflow>=2.13.0
# torch>=2.0.0
# torchvision>=0.15.0

# 3D Mesh & Geometry
trimesh>=3.20.0
pyvista>=0.42.0
networkx>=3.1
shapely>=2.0.0

# Web & Visualization
Flask>=2.3.0
flask-cors>=4.0.0

# Unity/Unreal Export & Validation
pydantic>=2.0.0

# Utilities
python-dotenv>=1.0.0
pillow>=10.0.0
tqdm>=4.66.0
jsonschema>=4.19.0

# Testing & Logging
pytest>=7.4.0
pytest-cov>=4.1.0

# Development
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0
